# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tax Exposure Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: companyId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: taxType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/upload:
    post:
      operationId: uploadTransactions
      tags: [transactions]
      summary: Upload transactions CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadTransactionsRequest"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        transactionDate:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        amount:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        accountCode:
          type: string
          nullable: true
        accountCategory:
          type: string
          nullable: true
        vendorName:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
        taxType:
          type: string
          nullable: true
        vatAmount:
          type: number
          nullable: true
        withholdingTaxAmount:
          type: number
          nullable: true
        transactionType:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - companyId
        - createdAt
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    UploadTransactionsRequest:
      type: object
      properties:
        companyId:
          type: string
        transactions:
          type: array
          items:
            type: object
            additionalProperties: true
        fileName:
          type: string
      required:
        - companyId
        - transactions
        - fileName
    UploadResult:
      type: object
      properties:
        uploadId:
          type: string
        rowsImported:
          type: integer
        message:
          type: string
      required:
        - uploadId
        - rowsImported
        - message
...